拆分页面布局
